plot_sensors method of SetChannelsMixin. Fix.
